+++ conv/conv.go
@@ -1,4 +1,4 @@
-package si
+package conv
 
 import "fmt"
